+++ crates/gem_ton/src/signer/mod.rs
@@ -5,9 +5,11 @@ mod signature;
 pub(crate) mod testkit;
 mod transaction;
 mod types;
+mod validator;
 
 pub use cells::{BagOfCells, BitReader, Cell, CellArc, CellBuilder};
 pub use chain_signer::TonChainSigner;
 pub use signature::sign_personal;
 pub use transaction::WalletV4R2;
 pub use types::{TonSignDataPayload, TonSignDataResponse, TonSignMessageData, TonSignResult};
+pub use validator::validate_network;
