+++ crates/gem_ton/src/signer/transaction/mod.rs
@@ -1,8 +1,6 @@
 mod message;
 mod request;
 mod signing;
-#[cfg(test)]
-mod tests;
 mod wallet;
 
 pub(crate) use signing::{sign_data, sign_swap, sign_token_transfer, sign_transfer};
